backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
requests>=2.31.0
requests-cache>=1.1.0
logbook>=1.7.0
python-dateutil>=2.8.2
roman>=3.3
pydantic>=2.0.0
cryptography>=42.0.0
pyyaml>=6.0.1
python-jose>=3.3.0
packaging>=23.2
